// frontend/src/pages/contracts/api/contracts.api.props.ts
export type UseAddContractProps = {
    onSuccess: () => void;
    onError?: (error: unknown) => void;
};

export type UseEditContractProps = {
    onSuccess: () => void;
    onError?: (error: unknown) => void;
};

export type UseDeleteContractProps = {
    onSuccess: () => void;
    onError?: (error: unknown) => void;
};